% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRFmet.R
\name{getModelDataDir}
\alias{getModelDataDir}
\title{Get Package Data Directory}
\usage{
getModelDataDir()
}
\value{
Absolute path string.
}
\description{
Returns the package data directory where model data is located.
}
\seealso{
\link{ModelDataDir}

\link{setModelDataDir}
}
