% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirFireWRF.R
\name{getWRFDataDir}
\alias{getWRFDataDir}
\title{Get package data directory}
\usage{
getWRFDataDir()
}
\value{
Absolute path string.
}
\description{
Returns the package data directory where model data is located.
}
\seealso{
\link{WRFDataDir}

\link{setWRFDataDir}
}
