% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_contours}
\alias{layer_contours}
\title{Create a contour lines layer for plotting}
\usage{
layer_contours(
  raster = NULL,
  varName = NULL,
  breaks = NULL,
  lineWidth = 0.5,
  color = "black",
  alpha = 1,
  ...
)
}
\arguments{
\item{raster}{A RasterBrick or RasterLayer}

\item{varName}{The name of a raster variable.}

\item{breaks}{A numerical vector of contour line levels.}

\item{lineWidth}{Width of contour lines.}

\item{color}{Color of contour lines.}

\item{alpha}{Transparency of layer.}

\item{...}{Additional parameters for ggplot2::geom_contour()}
}
\value{
A geom_contour ggproto object.
}
\description{
Create a contour lines layer for plotting
}
