% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_counties}
\alias{layer_counties}
\title{Create a plot layer for state county polygons}
\usage{
layer_counties(
  lineWidth = 0.5,
  color = "black",
  fill = "transparent",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{lineWidth}{Line width of county borders.}

\item{color}{Line color of county borders.}

\item{fill}{Fill color of county polygons.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}
}
\value{
A geom_polygon ggproto object.
}
\description{
Create a plot layer for state county polygons
}
