% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_raster}
\alias{layer_raster}
\title{Create a plot layer for a raster}
\usage{
layer_raster(raster = NULL, varName = NULL, breaks = NULL, alpha = 1)
}
\arguments{
\item{raster}{A RasterBrick or RasterLayer.}

\item{varName}{The name of a raster variable.}

\item{breaks}{A vector of raster values to use as palette breaks.}

\item{alpha}{Transparency of layer.}
}
\value{
A geom_tile ggproto object.
}
\description{
Create a plot layer for a raster
}
