% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_spPolys}
\alias{layer_spPolys}
\title{Create a spatial polygons layer for plotting}
\usage{
layer_spPolys(spdf = NULL, color = "black", fill = "transparent")
}
\arguments{
\item{spdf}{A SpatialPolygonsDataFrame.}

\item{color}{Outline color.}

\item{fill}{Fill color.}
}
\value{
A geom_polygon ggproto object.
}
\description{
Create a spatial polygons layer for plotting
}
