% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_vectorField}
\alias{layer_vectorField}
\title{Create a plot layer for a vector field}
\usage{
layer_vectorField(
  raster = NULL,
  uName = NULL,
  vName = NULL,
  arrowCount = 1000,
  arrowScale = 0.05,
  arrowHead = 0.05,
  arrowWidth = 1.2,
  arrowColor = "dodgerblue",
  alpha = 1,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{raster}{A RasterBrick with layers for u/v vector components.}

\item{uName}{The name of the u component layer.}

\item{vName}{The name of the v component layer.}

\item{arrowCount}{Number of arrows to draw.}

\item{arrowScale}{Scale factor of arrow length.}

\item{arrowHead}{Size of arrow head.}

\item{arrowWidth}{Line width of arrows.}

\item{arrowColor}{Color of arrows.}

\item{alpha}{Transparency of layer.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}
}
\value{
An annotation_custom ggproto object.
}
\description{
Create a plot layer for a vector field
}
