% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRFmet.R
\docType{data}
\name{ModelDataDir}
\alias{ModelDataDir}
\title{Directory for Modeling Data}
\format{
Absolute path string.
}
\description{
This package maintains an internal directory location which
users can set using \code{setModelDataDir()}. All package functions use this
directory whenever datasets are created or loaded.

The default setting when the package is loaded is \code{getwd()}.
}
\seealso{
\link{getModelDataDir}

\link{setModelDataDir}
}
\keyword{environment}
