% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_base}
\alias{plot_base}
\title{Creates an empty plot}
\usage{
plot_base(
  title = NULL,
  xlab = "Longitude",
  ylab = "Latitude",
  color = NULL,
  flab = NULL,
  xlim = NULL,
  ylim = NULL,
  ratio = NULL
)
}
\arguments{
\item{title}{Title of the plot.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{color}{Label for color value in legend.}

\item{flab}{Label for fill value in legend.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{ratio}{Plot dimensions ratio (x/y).}
}
\value{
A ggplot object.
}
\description{
Creates an empty plot
}
