% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_raster}
\alias{plot_raster}
\title{Creates a plot for a single raster layer}
\usage{
plot_raster(
  raster = NULL,
  fillLow = "#132c43",
  fillHigh = "#55b2f8",
  fillNa = "transparent",
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  flab = NULL,
  xlim = NULL,
  ylim = NULL,
  ratio = NULL
)
}
\arguments{
\item{raster}{A RasterLayer.}

\item{fillLow}{Color for lowest raster value.}

\item{fillHigh}{Color for highest raster value.}

\item{fillNa}{Color for na raster values.}

\item{title}{Title of plot.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{flab}{Label for legend fill scale.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{ratio}{Aspect ratio of plot.}
}
\value{
A ggplot object.
}
\description{
Creates a plot for a single raster layer
}
\examples{
\donttest{
library(WRFmet)
library(raster)

setWRFDataDir("~/Data/WRF")

plot_raster(
  raster = example_PNW$HGT,
  title = "PNW Elevation",
  flab = "Meters",
  ratio = 1.4
)
}
}
