% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRFmet.R
\docType{data}
\name{PNW_Sample}
\alias{PNW_Sample}
\title{PNW Example Dataset}
\format{
A RasterBrick with four layers
}
\description{
This dataset contains forecasted readings for temperature (TSK), elevation 
(HGT), and wind velocity (U10 & V10) across the Pacific Northwest on July 15,
2020. Each variable is stored in its own \emph{RasterLayer} in an overlying
\emph{RasterBrick} object. This data is used in some examples in the package
documentation.
}
\details{
PNW-4km example dataset
}
\keyword{datasets}
