% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRFmet.R
\name{removeModelDataDir}
\alias{removeModelDataDir}
\title{Remove package data directory}
\usage{
removeModelDataDir()
}
\value{
Silently returns previous value of data directory.
}
\description{
Resets the package data directory to NULL. Used for internal
testing.
}
\seealso{
ModelDataDir

getModelDataDir

setModelDataDir
}
\keyword{internal}
