% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRFmet.R
\name{removeWRFDataDir}
\alias{removeWRFDataDir}
\title{Remove package data directory}
\usage{
removeWRFDataDir()
}
\value{
Silently returns previous value of data directory.
}
\description{
Resets the package data directory to NULL. Used for internal
testing.
}
\seealso{
WRFDataDir

getWRFDataDir

setWRFDataDir
}
\keyword{internal}
