% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRFmet.R
\name{setModelDataDir}
\alias{setModelDataDir}
\title{Set Package Data Directory}
\usage{
setModelDataDir(dataDir)
}
\arguments{
\item{dataDir}{directory where model datasets are located}
}
\value{
Silently returns previous value of data directory.
}
\description{
Sets the package data directory where model data is located.
If the directory does not exist, it will be created.
}
\seealso{
\link{ModelDataDir}

\link{getModelDataDir}
}
