% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRFmet.R
\name{setWRFDataDir}
\alias{setWRFDataDir}
\title{Set package data directory}
\usage{
setWRFDataDir(dataDir)
}
\arguments{
\item{dataDir}{directory where model datasets are located}
}
\value{
Silently returns previous value of data directory.
}
\description{
Sets the package data directory where model data is located.
If the directory does not exist, it will be created.
}
\seealso{
\link{WRFDataDir}

\link{getWRFDataDir}
}
