% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_download.R
\name{wrf_download}
\alias{wrf_download}
\title{Download WRf model data from AirFire}
\usage{
wrf_download(
  modelName = NULL,
  modelRun = NULL,
  modelRunHour = NULL,
  baseUrl = "http://m2.airfire.org/",
  verbose = TRUE
)
}
\arguments{
\item{modelName}{Model identifier.}

\item{modelRun}{Model initialization timestamp as "YYYYMMDDHH".}

\item{modelRunHour}{Hour forecasted from initial time, i.e. 7.}

\item{baseUrl}{Base URL for WRF output.}

\item{verbose}{If \code{FALSE}, suppress status messages (if any), and the
progress bar.}
}
\description{
Downloads a copy of the specified WRF model run to the package 
data directory. The file data can then be loaded with \code{wrf_load()}.

On 2020-08-26, available model identifiers include the following:
\itemize{
  \item{PNW-1.33km}
  \item{PNW-4km}
}
}
\examples{
\donttest{
library(WRFmet)
setWRFDataDir('~/Data/WRF')

modelRun <- wrf_latestModelRun("PNW-4km")
wrf_download("PNW-4km", "2020082500", 11)
}
}
