% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_downloaded.R
\name{wrf_downloaded}
\alias{wrf_downloaded}
\title{List downloaded WRF model run files}
\usage{
wrf_downloaded(
  WRFDataDir = getWRFDataDir(),
  pattern = ".nc",
  full = FALSE,
  ...
)
}
\arguments{
\item{WRFDataDir}{Directory in which model run files are located.}

\item{pattern}{A regex pattern to use for filtering model files.}

\item{full}{Logical. Show the full path of the model (used for local loading).}

\item{...}{Additional arguments to be passed to \code{list.files()}.}
}
\value{
A list of downloaded model run files
}
\description{
List downloaded WRF model run files
}
\examples{
\donttest{
library(AirFireWRF)
setWRFDataDir("~/Data/WRF")

modelRun <- wrf_latestModelRun("PNW-4km")
filePath <- wrf_download("PNW-4km", modelRun, 7)
wrf_downloaded()
}
}
