% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_listModelRuns.R
\name{wrf_listLatestModelRun}
\alias{wrf_listLatestModelRun}
\title{List latest WRF model run available from AirFire}
\usage{
wrf_listLatestModelRun(
  modelName = "PNW-4km",
  baseUrl = "http://m2.airfire.org/PNW/4km/WRF/",
  count = 1
)
}
\arguments{
\item{modelName}{Model identifier.}

\item{baseUrl}{Base URL for WRF output.}

\item{count}{Number of most recent model run strings to return.}
}
\value{
The latest model run string.
}
\description{
Finds the latest WRF model runs available from AirFire and lists
the most recent \code{count} model run strings.
}
\examples{
\donttest{
library(WRFmet)
wrf_listLatestModelRun(count = 2)
}
}
