% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_listModelRuns.R
\name{wrf_listModelRuns}
\alias{wrf_listModelRuns}
\title{List WRF model runs available from AirFire}
\usage{
wrf_listModelRuns(
  modelName = "PNW-4km",
  baseUrl = "http://m2.airfire.org/PNW/4km/WRF/"
)
}
\arguments{
\item{modelName}{Model identifier.}

\item{baseUrl}{Base URL for WRF output.}
}
\value{
A list of WRF model run strings.
}
\description{
Finds all of the WRF model runs available from AirFire.
}
\examples{
\donttest{
library(WRFmet)
wrf_listModelRuns()
}
}
