% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_load.R
\name{wrf_load}
\alias{wrf_load}
\title{Load data for a single WRF model}
\usage{
wrf_load(localPath = NULL, xlim = NULL, ylim = NULL, verbose = TRUE)
}
\arguments{
\item{localPath}{Absolute path to a NetCDF file not found in `modelDataDir`.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{verbose}{Logical to display messages.}
}
\value{
A \pkg{raster} package \emph{RasterBrick} object.
}
\description{
Load data for a single WRF model
}
\examples{
\donttest{
library(WRFmet)
library(ggplot2)
library(maps)

localPath <- "~/Data/WRF/wrfout_d3-2020071512-f07-0000.nc"
raster <- wrf_load(localPath)

states <- map_data("state")

rasterVis::gplot(raster) +
  geom_raster(aes(fill = .data$value)) +
  geom_polygon(data = states, aes(x = long, y = lat, group = group), color = "black", fill = NA) +
  scale_fill_continuous(na.value = "red")
}
}
