% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_loadTimeSeries.R
\name{wrf_loadTimeSeries}
\alias{wrf_loadTimeSeries}
\title{Load time series WRF model run data}
\usage{
wrf_loadTimeSeries(
  modelName = NULL,
  modelRun = NULL,
  modelRunHours = NULL,
  baseUrl = "http://m2.airfire.org",
  varName = NULL,
  res = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{modelName}{Model identifier.}

\item{modelRun}{Model initialization timestamp as "YYYYMMDDHH".}

\item{modelRunHours}{Vector of hours forecasted from init time, i.e. 7:24.}

\item{baseUrl}{Base URL for WRF output.}

\item{varName}{WRF variable to load.}

\item{res}{Resolution of raster in degrees.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}
}
\value{
A \pkg{raster} package \emph{RasterBrick} object.
}
\description{
Loads time series WRF model run data as a \pkg{raster} package 
\emph{RasterBrick}. If the specified data does not exist on the user's 
machine, then it will try to be downloaded from the AirFire database.

On 2020-09-17, available model identifiers include the following:
\itemize{
  \item{PNW-1.33km}
  \item{PNW-4km}
}
}
\examples{
\donttest{
# library(WRFmet)
# setWRFDataDir("~/Data/WRF")

#tempTimeSeries <- wrf_loadTimeSeries(
# modelName = "PNW-1.33km",
# modelRun = "2020091712",
# modelRunHours = 7:10,
# varName = "TSK",
# res = 0.05,
#  xlim = c(-125, -117),
# ylim = c(45, 49)
# )

# rasterVis::levelplot(tempTimeSeries)
}
}
