% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_plots.R
\name{wrf_standardPlot}
\alias{wrf_standardPlot}
\title{Creates a comprehensive plot}
\usage{
wrf_standardPlot(
  raster = NULL,
  states = FALSE,
  polys = NULL,
  bgName = NULL,
  ctrName = NULL,
  uName = NULL,
  vName = NULL,
  bgRasterColors = grDevices::terrain.colors(10),
  bgRasterValues = NULL,
  bgRasterNaColor = "transparent",
  stateWidth = 0.5,
  stateColor = "red",
  stateFill = "transparent",
  polyWidth = 0.5,
  polyColor = "red",
  polyFill = "transparent",
  ctrBreaks = NULL,
  ctrWidth = 0.25,
  ctrColor = "black",
  arrowCount = 1000,
  arrowScale = 0.05,
  arrowWidth = 0.8,
  arrowColor = "black",
  arrowHead = 0.05,
  arrowAlpha = 1,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  flab = NULL,
  xlim = NULL,
  ylim = NULL,
  ratio = NULL
)
}
\arguments{
\item{raster}{A RasterBrick with layers for WRF variables.}

\item{states}{Logical for including state polygons or not.}

\item{polys}{A SpatialPolygonsDataFrame.}

\item{bgName}{The name of the background raster layer.}

\item{ctrName}{The name of the contour raster layer.}

\item{uName}{The name of the u component raster layer.}

\item{vName}{The name of the v component raster layer.}

\item{bgRasterColors}{Vector of colors to use for the background raster's 
n-color gradient.}

\item{bgRasterValues}{A vector of positions (between 0 and 1) for each color 
in the bgRasterColors vector.}

\item{bgRasterNaColor}{Color for na background raster values.}

\item{stateWidth}{Line thickness of state polygon outlines.}

\item{stateColor}{Color for state polygon outlines.}

\item{stateFill}{Color for state polygon interiors.}

\item{polyWidth}{Line thickness of spatial polygon outlines.}

\item{polyColor}{Color for spatial polygon outlines.}

\item{polyFill}{Color for spatial polygon interiors.}

\item{arrowCount}{Number of vector field arrows to draw.}

\item{arrowScale}{Scale factor of vector field arrow body length.}

\item{arrowWidth}{Line thickness of vector field arrows.}

\item{arrowColor}{Color of vector field arrows.}

\item{arrowHead}{Size of vector field arrowheads.}

\item{arrowAlpha}{Transparency of vector field arrows.}

\item{title}{Title of plot.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{flab}{Label for legend fill scale.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{ratio}{Aspect ratio of plot.}
}
\value{
A ggplot object.
}
\description{
Creates a comprehensive plot
}
\examples{
\donttest{
library(WRFmet)
library(raster)

wrf_standardPlot(
  raster = example_PNW,
  bgName = "HGT",
  uName = "U10",
  vName = "V10",
  states = TRUE,
  arrowAlpha = 0.75,
  title = "PNW Elevation & Wind Velocity",
  flab = "Elev (m)",
  xlim = c(-125, -111),
  ylim = c(42, 49),
  ratio = 1.4
)
}
}
