% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_windField.R
\name{wrf_windField}
\alias{wrf_windField}
\title{Plot a vectorfield of wind velocity}
\usage{
wrf_windField(
  nc = NULL,
  res = NULL,
  xlim = NULL,
  ylim = NULL,
  bg = "HGT",
  arrowCount = 1000,
  arrowScale = 1,
  arrowColor = "white",
  headSize = 0.05
)
}
\arguments{
\item{nc}{WRF NetCDF file.}

\item{res}{Resolution of raster in degrees.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{bg}{A name of a WRF variable to use as a background raster or FALSE for
no background.}

\item{arrowCount}{The number of arrows to plot.}

\item{arrowScale}{Numeric multiplier for arrow body length.}

\item{arrowColor}{Color of arrows.}

\item{headSize}{Size of arrowhead.}
}
\value{
A \emph{Trellis} object.
}
\description{
Plot a vectorfield of wind velocity
}
\examples{
\donttest{
library(WRFmet)
nc <- ncdf4::nc_open("~/Data/WRF/wrfout_d3-2020071512-f07-0000.nc")
wrf_windField(
  nc = nc,
  res = 0.08,
  xlim = c(-126, -115),
  ylim = c(45, 51),
  bg = "HGT",
  arrowScale = 0.05,
  arrowColor = 'white',
  headSize = 0.05
)
}
}
