% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRFmet.R
\docType{data}
\name{WRFDataDir}
\alias{WRFDataDir}
\title{Directory for modeling data}
\format{
Absolute path string.
}
\description{
This package maintains an internal directory location which
users can set using \code{setWRFDataDir()}. All package functions use this
directory whenever datasets are created or loaded.

The default setting when the package is loaded is \code{getwd()}.
}
\seealso{
\link{getWRFDataDir}

\link{setWRFDataDir}
}
\keyword{environment}
