% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_points}
\alias{layer_points}
\title{Create a points layer for plotting}
\usage{
layer_points(points = NULL, size = 1)
}
\arguments{
\item{points}{A SpatialPointsDataFrame of coordinates and data.}

\item{size}{Point size.}
}
\value{
A geom_point ggproto object.
}
\description{
Create a points layer for plotting
}
