% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_raster}
\alias{layer_raster}
\title{Create a Raster layer for plotting}
\usage{
layer_raster(raster = NULL, alpha = 1)
}
\arguments{
\item{raster}{A RasterLayer or a RasterBrick with one layer.}

\item{alpha}{Transparency of layer.}
}
\value{
A geom_tile ggproto object.
}
\description{
Create a Raster layer for plotting
}
