% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_spPolys}
\alias{layer_spPolys}
\title{Create a plot layer for spatial polygons}
\usage{
layer_spPolys(
  polygons = NULL,
  lineWidth = 0.5,
  color = "black",
  fill = "transparent"
)
}
\arguments{
\item{polygons}{A SpatialPolygonsDataFrame.}

\item{lineWidth}{Line width of polygon outlines.}

\item{color}{Outline color.}

\item{fill}{Fill color.}
}
\value{
A geom_polygon ggproto object.
}
\description{
Create a plot layer for spatial polygons
}
