% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_states}
\alias{layer_states}
\title{Create a state polygons layer for plotting}
\usage{
layer_states(
  lineWidth = 0.5,
  color = "black",
  fill = "transparent",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{lineWidth}{Line width of state borders.}

\item{color}{Line color of state borders.}

\item{fill}{Fill color of state polygons.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}
}
\value{
A geom_polygon ggproto object.
}
\description{
Create a state polygons layer for plotting
}
