% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_vectorField}
\alias{layer_vectorField}
\title{Create a vectorfield layer for plotting}
\usage{
layer_vectorField(
  uvRaster = NULL,
  xlim = NULL,
  ylim = NULL,
  arrowCount = 1000,
  arrowScale = 0.05,
  arrowColor = "white",
  headSize = 0.05,
  alpha = 1
)
}
\arguments{
\item{uvRaster}{A RasterBrick with 2 RasterLayers: U and V vector components.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{arrowCount}{Number of arrows to draw.}

\item{arrowScale}{Arrow length scale factor.}

\item{arrowColor}{Arrow color.}

\item{headSize}{Arrow head size.}

\item{alpha}{Transparency of layer.}
}
\value{
An annotation_custom ggproto object.
}
\description{
Create a vectorfield layer for plotting
}
