% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_base}
\alias{plot_base}
\title{Creates an empty plot}
\usage{
plot_base(
  title = NULL,
  xlab = "Longitude",
  ylab = "Latitude",
  clab = NULL,
  flab = NULL,
  xlim = NULL,
  ylim = NULL,
  ratio = NULL
)
}
\arguments{
\item{title}{Title of plot.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{clab}{Label for legend color scale.}

\item{flab}{Label for legend fill scale.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{ratio}{Aspect ratio of plot.}
}
\value{
A ggplot object.
}
\description{
Creates an empty plot
}
\keyword{internal}
