% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_raster}
\alias{plot_raster}
\title{Creates a plot for a single raster layer}
\usage{
plot_raster(
  ...,
  raster = NULL,
  colors = grDevices::terrain.colors(10),
  values = NULL,
  fillNa = "transparent",
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  flab = NULL,
  xlim = NULL,
  ylim = NULL,
  ratio = NULL
)
}
\arguments{
\item{...}{Arguments passed on to ggplot2::continuous_scale()}

\item{raster}{A RasterLayer.}

\item{colors}{A vector of colours to use for n-colour gradient.}

\item{values}{A vector of positions (between 0 and 1) for each color in the 
colors vector.}

\item{fillNa}{Color for na raster values.}

\item{title}{Title of plot.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{flab}{Label for legend fill scale.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{ratio}{Aspect ratio of plot.}
}
\value{
A ggplot object.
}
\description{
Creates a plot for a single raster layer
}
\examples{
\donttest{
library(WRFmet)
library(raster)

plot_raster(
  raster = example_PNW$HGT,
  title = "PNW Elevation",
  flab = "Meters",
  ratio = 1.4
)
}
}
