% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_raster}
\alias{plot_raster}
\title{Creates a plot for a single raster layer}
\usage{
plot_raster(
  raster = NULL,
  fillLow = "#132c43",
  fillHigh = "#55b2f8",
  fillNa = "transparent",
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  flab = NULL,
  xlim = NULL,
  ylim = NULL,
  ratio = NULL
)
}
\arguments{
\item{raster}{A RasterLayer.}

\item{fillLow}{Color for lowest raster value.}

\item{fillHigh}{Color for highest raster value.}

\item{fillNa}{Color for na raster values.}

\item{title}{Title of the plot.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{flab}{Label for fill value in legend.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{ratio}{Plot dimensions ratio (x/y).}
}
\value{
A GGPlot object.
}
\description{
Creates a plot for a single raster layer
}
\examples{
\donttest{
library(WRFmet)

plot_raster(
  raster = sample_PNW$HGT,
  ratio = 1.4
)
}
}
