% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_raster}
\alias{plot_raster}
\title{Creates a plot for a single raster layer}
\usage{
plot_raster(raster = NULL)
}
\arguments{
\item{raster}{A RasterLayer.}
}
\value{
A ggplot object.
}
\description{
Creates a plot for a single raster layer
}
