% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_standard}
\alias{plot_standard}
\title{Creates a comprehensive plot}
\usage{
plot_standard(
  ...,
  bgRaster = NULL,
  states = FALSE,
  polys = NULL,
  uRaster = NULL,
  vRaster = NULL,
  bgRasterColors = grDevices::terrain.colors(10),
  bgRasterValues = NULL,
  bgRasterNaColor = "transparent",
  stateColor = "black",
  stateFill = "white",
  polyColor = "black",
  polyFill = "white",
  arrowCount = 1000,
  arrowScale = 0.05,
  arrowColor = "white",
  arrowHead = 0.05,
  arrowAlpha = 1,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  flab = NULL,
  xlim = NULL,
  ylim = NULL,
  ratio = NULL
)
}
\arguments{
\item{...}{Arguments passed on to ggplot2::continuous_scale()}

\item{bgRaster}{A RasterLayer for the background.}

\item{states}{Logical for including state polygons or not.}

\item{polys}{A SpatialPolygonsDataFrame.}

\item{uRaster}{A RasterLayer for longitudinal vector components.}

\item{vRaster}{A RasterLayer for latitudinal vector components.}

\item{bgRasterColors}{Vector of colours to use for the bgRaster's n-colour 
gradient.}

\item{bgRasterValues}{A vector of positions (between 0 and 1) for each color in the 
colors vector.}

\item{bgRasterNaColor}{Color for na ngRaster values.}

\item{stateColor}{Color for state polygon outlines.}

\item{stateFill}{Color for state polygon interiors.}

\item{polyColor}{Color for spatial polygon outlines.}

\item{polyFill}{Color for spatial polygon interiors.}

\item{arrowCount}{Number of arrows to draw.}

\item{arrowScale}{Arrow length scale factor.}

\item{arrowColor}{Arrow color.}

\item{arrowHead}{Arrow head size.}

\item{arrowAlpha}{Transparency of vector field layer.}

\item{title}{Title of plot.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{flab}{Label for legend fill scale.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{ratio}{Aspect ratio of plot.}
}
\value{
A ggplot object.
}
\description{
Creates a comprehensive plot
}
\examples{
\donttest{
library(WRFmet)
library(raster)

plot_standard(
  bgRaster = example_PNW$HGT,
  uRaster = example_PNW$U10,
  vRaster = example_PNW$V10,
  states = TRUE,
  stateColor = "red",
  stateFill = "transparent",
  arrowColor = "black",
  title = "PNW-4km 2020-07-15 12pm - Hour 7",
  flab = "Elev (m)",
  xlim = c(-125, -111),
  ylim = c(42, 49),
  ratio = 1.4
)
}
}
