% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_createRaster.R
\name{wrf_createRaster}
\alias{wrf_createRaster}
\title{Load rasterized data for a WRF model

Some useful variables:
\itemize{
  \item{XLAT}
  \item{XLONG}
  \item{XLAT_U}   
  \item{XLAT_V}
  \item{XLONG_U}
  \item{XLONG_V}
  \item{U - Wind X component (m/s)}
  \item{V - Wind Y component (m/s)}
  \item{W - Wind Z component (m/s)}
  \item{U10 - Wind X component at 10 meters (m/s)}
  \item{V10 - Wind Y component at 10 meters (m/s)}
  \item{ZNU}
  \item{ZNW}
  \item{LU_INDEX - Land use category}
  \item{Q2}
  \item{T - Perturbation potential temperature theta-t0 (K)}
  \item{T2 - Temperature at 2 meters (K)}
  \item{TH2 - Potential temperature at 2 meters (K)}
  \item{HGT - Terrain height (m)}
  \item{RAINNC - Accumulated total grid scale precipitation (mm)}
  \item{CFRACT - Total cloud fraction}
  \item{PBLH - Planetary boundary layer height (m)}
}}
\usage{
wrf_createRaster(
  nc = NULL,
  vars = NULL,
  res = NULL,
  xlim = NULL,
  ylim = NULL,
  verbose = TRUE
)
}
\arguments{
\item{nc}{WRF NetCDF file.}

\item{vars}{WRF variable(s) to load.}

\item{res}{Resolution of raster in degrees.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{verbose}{Logical to display messages.}
}
\value{
A \pkg{raster} package \emph{RasterBrick} object.
}
\description{
Load rasterized data for a WRF model

Some useful variables:
\itemize{
  \item{XLAT}
  \item{XLONG}
  \item{XLAT_U}   
  \item{XLAT_V}
  \item{XLONG_U}
  \item{XLONG_V}
  \item{U - Wind X component (m/s)}
  \item{V - Wind Y component (m/s)}
  \item{W - Wind Z component (m/s)}
  \item{U10 - Wind X component at 10 meters (m/s)}
  \item{V10 - Wind Y component at 10 meters (m/s)}
  \item{ZNU}
  \item{ZNW}
  \item{LU_INDEX - Land use category}
  \item{Q2}
  \item{T - Perturbation potential temperature theta-t0 (K)}
  \item{T2 - Temperature at 2 meters (K)}
  \item{TH2 - Potential temperature at 2 meters (K)}
  \item{HGT - Terrain height (m)}
  \item{RAINNC - Accumulated total grid scale precipitation (mm)}
  \item{CFRACT - Total cloud fraction}
  \item{PBLH - Planetary boundary layer height (m)}
}
}
\examples{
\donttest{
library(WRFmet)
nc <- ncdf4::nc_open("~/Data/WRF/wrfout_d3-2020071512-f07-0000.nc")
raster <- wrf_createRaster(
  nc = nc,
  vars = c("HGT", "TSK"),
  res = 0.1,
  xlim = c(-125, -116),
  ylim = c(45, 50)
)
print(raster)
}
}
