% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_createRaster.R
\name{wrf_createRaster}
\alias{wrf_createRaster}
\title{Load rasterized data for a WRF model}
\usage{
wrf_createRaster(
  nc = NULL,
  vars = NULL,
  xlim = NULL,
  ylim = NULL,
  verbose = TRUE
)
}
\arguments{
\item{nc}{WRF NetCDF file.}

\item{vars}{WRF variable(s) to load.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{verbose}{Logical to display messages.}
}
\value{
A \pkg{raster} package \emph{RasterBrick} object.
}
\description{
Load rasterized data for a WRF model
}
\examples{
\donttest{
library(WRFmet)
nc <- ncdf4::nc_open("~/Data/WRF/wrfout_d3-2020071512-f07-0000.nc")
raster <- wrf_createRaster(nc, c("HGT", "TSK"))
print(raster)
}
}
