% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_plots.R
\name{wrf_rasterPlot}
\alias{wrf_rasterPlot}
\title{Creates a plot for a single raster layer}
\usage{
wrf_rasterPlot(
  raster = NULL,
  varName = NULL,
  breaks = NULL,
  palette = "Blues",
  paletteDir = 1,
  colors = grDevices::terrain.colors(10),
  values = NULL,
  naColor = "transparent",
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  flab = NULL,
  xlim = NULL,
  ylim = NULL,
  ratio = NULL
)
}
\arguments{
\item{raster}{A RasterBrick or RasterLayer.}

\item{varName}{The name of a raster variable.}

\item{breaks}{A vector of raster values to use as palette breaks.}

\item{palette}{The name of a Brewer palette.}

\item{paletteDir}{Direction of the Brewer palette (1 = standard, -1 = reverse)}

\item{colors}{A vector of colors to use for n-color gradient.}

\item{values}{A vector of positions (between 0 and 1) for each color in the 
colors vector.}

\item{naColor}{Color for na raster values.}

\item{title}{Title of plot.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{flab}{Label for legend fill scale.}

\item{xlim}{A vector of coordinate longitude bounds.}

\item{ylim}{A vector of coordinate latitude bounds.}

\item{ratio}{Aspect ratio of plot.}
}
\value{
A ggplot object.
}
\description{
Creates a plot for a single raster layer
}
\details{
The raster defaults to a gradient fill color scale when 
\code{breaks} is not specified. Otherwise, it will be colored using a 
discrete Brewer palette.
}
\examples{
\donttest{
library(WRFmet)
library(raster)

wrf_rasterPlot(
  raster = example_PNW,
  varName = "HGT",
  flab = "Elev (m)",
  ratio = 1.4
)
}
}
