% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRFmet.R
\docType{package}
\name{WRFmet}
\alias{WRFmet}
\title{Utilities for working with WRF model output}
\description{
A suite of utility functions providing functions commonly
required for working with WRF model output in R.
}
